requirements.txt
Flask==3.1.0
flask_wtf==1.2.2
moviepy==1.0.3
openai_whisper==20240930
scipy==1.15.2
torch==2.5.1
transformers==4.48.0
word2number==1.1
WTForms==3.2.1
